Moved index.html so we can check for authentication and serve ourselves from server.js.
Login, logout flow working with olinapps. Visitors may not access the bingo app unless they are authenticated.

Olinapps authentication is connected to database. Users are verified to persist.

Users' names are not capitalized. I've added a fxn in the new 'utils.js' file to take care of this. This will be tested and added next commit.